moved event handler to game.py
menus.py now only checks mouse collisions. This should allow for more
advanced processes to occur from mouse interactions such as a exit
button.

Bug fix: checking wrong values when checking mouse_state as mouse state
contains a tuple of integers used to represent True and False. This
means that the program was testing to see if 1 was True when it was only
equivalent

Signed-off-by: xedre <git@xed.re>